(building) Accept sprites without symmetries.
This introduces a behavior change depending on the type of the argument
of methods, which is annoying to have, and I still think it is better if
we can do without this.

That said, making night masks of sprites without symmetries has obvious
uses *right now* (purchase preview of station templates), so I have to
get things working first.