(third-party) Import more OpenGFX foundations.
These are all for the temperate climate. I'll add their counterparts for
other climates when needed.

The foundation supporting code is mostly non-existent at this moment,
and I am not sure I want to implement these, since their purpose for now
is just to make those station demos, which uses a small subset of all
the possible combinations.