(building) Foundation refactor.
Instead of asking for a grf.AlternativeSprites as the basis for
foundation sprites, we make a new dataclass that contains all relevant
information for creating a foundation. This should be more flexible.

The implementation still has missing parts and (probably) bugs. However,
since this breaks interface I have to merge it as soon as possible.